🚸 update: 登録確認中のローディングを追加
